refactor: Update fetchEspresso function to use model.HttpCustomError
